requirements.txt
Django==1.4.2
PIL==1.1.7
South==0.7.6
django-daguerre==0.1
wsgiref==0.1.2
-e src/bees
